refactor(policy-creator): rename rego-editor.tsx → code-editor.tsx
The file primarily exports CodeEditor (a language-agnostic editor)
and only secondarily exports RegoEditor as a backwards-compat
alias. Naming the module after the secondary export was confusing
review and grep, so flip the filename to match the primary export
and update both call sites (live-test.tsx + custom-rego.tsx) to
import from the new path. RegoEditor stays as a thin shim so
existing Rego-only call sites don't have to thread defaults.

No behavior change.
